modified:   ../../ainsuForthsketch.cpp 	new file:   __ORIG_fatfs_full_usage.ino__ 	new file:   fatfs.cpp 	new file:   fatfs.h
Provides a dump of main.py if present on the SPI flash.  No error checking.

For some reason, there's only one SPI interface (there are
two on the Metro M0 Express and also on the Feather M0 Express.

So the example code (from the Adafruit_SPIFlash library) did
not want to run (everything had a '1' after it: SPI became SPI1,
and so on).  Removed two of those '1' instances where the compiler
complains (beginning with SS1 -- SPI1 was the other one, iirc).

IN THE END, the local copy (here: fatfs.cpp) was modified, not
the corresponding board library -- did not modify the files that
were found deep under:

   ~/.arduino15/packages/adafruit/hardware/samd/1.0.18